seperating functions
I have seperated the class and funcs into their own file.
created a file that should be run a main.
thus far it is working.
