--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "C:\\ProgramData\\Anaconda3\\envs\\musicPlayer\\python.exe"
-}
